requirements.txt
accelerate==1.12.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
attrs==25.4.0
beautifulsoup4==4.14.3
boto3==1.42.30
botocore==1.42.30
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cobble==0.1.4
coloredlogs==15.0.1
colorlog==6.10.1
cryptography==46.0.3
defusedxml==0.7.1
dill==0.4.0
docling==2.68.0
docling-core==2.59.0
docling-ibm-models==3.10.3
docling-parse==4.7.3
elastic-transport==9.2.1
elasticsearch==9.2.1
et_xmlfile==2.0.0
Faker==40.1.2
fastapi==0.104.1
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-core==1.2.7
langchain-text-splitters==1.1.0
langsmith==0.6.4
latex2mathml==3.78.1
lxml==6.0.2
magika==0.6.3
mammoth==1.11.0
markdown-it-py==4.0.0
markdownify==1.2.2
markitdown==0.1.4
marko==2.2.2
MarkupSafe==3.0.3
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
networkx==3.6.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
onnxruntime==1.23.2
opencv-python==4.12.0.88
openpyxl==3.1.5
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==11.3.0
pluggy==1.6.0
polyfactory==3.2.0
protobuf==6.33.4
psutil==7.2.1
pyclipper==1.4.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pylatexenc==2.10
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
pypdfium2==4.30.0
pytesseract==0.3.10
pytest==7.4.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
rapidocr==3.5.0
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.9
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rtree==1.4.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
semchunk==2.2.2
sentence-transformers==5.2.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
starlette==0.27.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.6
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
triton==3.5.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.0.7
uuid_utils==0.13.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wheel==0.45.1
xlsxwriter==3.2.9
zstandard==0.25.0


# Google Drive API
google-api-python-client==2.157.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
